requirements.txt
pulpcore-client
pulp_file-client
ansible>=2.9
ansible_runner
flake8
pylint
pycodestyle
voluptuous
yamllint
pytest
vcrpy
